extract reader / writer #7
