+++ src/miniFasta/__init__.py
@@ -1,5 +1,6 @@
-from ._miniFasta import fasta_object, read, write, print_fasta, translate_seq, reverse_comp
-
+from ._miniFasta import print_fasta, translate_seq, reverse_comp
+from ._reader import read
+from ._writer import write, fasta_object
 __all__ = [
     "fasta_object",
     "read",
